+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/protocol/datalake_rest_client.hpp
@@ -91,7 +91,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     constexpr static const char* HeaderXMsLeaseBreakPeriod = "x-ms-lease-break-period";
     constexpr static const char* HeaderLeaseTime = "x-ms-lease-time";
     constexpr static const char* HeaderAcceptRanges = "accept-ranges";
-    constexpr static const char* HeaderContentRange = "content-range";
     constexpr static const char* HeaderResourceType = "x-ms-resource-type";
     constexpr static const char* HeaderLeaseState = "x-ms-lease-state";
     constexpr static const char* HeaderLeaseStatus = "x-ms-lease-status";
@@ -404,8 +403,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       Azure::Core::Nullable<std::string> AcceptRanges;
       PathHttpHeaders HttpHeaders;
-      Azure::Core::Http::Range ContentRange;
-      int64_t FileSize;
       std::string ETag;
       Core::DateTime LastModified;
       Azure::Core::Nullable<std::string> ResourceType;
@@ -1913,36 +1910,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
             if (response.GetHeaders().find("content-language") != response.GetHeaders().end())
             {
               result.HttpHeaders.ContentLanguage = response.GetHeaders().at("content-language");
-            }
-
-            auto content_range_iterator = response.GetHeaders().find(Details::HeaderContentRange);
-            if (content_range_iterator != response.GetHeaders().end())
-            {
-              const std::string& content_range = content_range_iterator->second;
-              auto bytes_pos = content_range.find("bytes ");
-              auto dash_pos = content_range.find("-", bytes_pos + 6);
-              auto slash_pos = content_range.find("/", dash_pos + 1);
-              int64_t range_start_offset = std::stoll(std::string(
-                  content_range.begin() + bytes_pos + 6, content_range.begin() + dash_pos));
-              int64_t range_end_offset = std::stoll(std::string(
-                  content_range.begin() + dash_pos + 1, content_range.begin() + slash_pos));
-              result.ContentRange = Azure::Core::Http::Range{
-                  range_start_offset, range_end_offset - range_start_offset + 1};
-            }
-            else
-            {
-              result.ContentRange = Azure::Core::Http::Range{
-                  0, std::stoll(response.GetHeaders().at(Details::HeaderContentLength))};
-            }
-            if (content_range_iterator != response.GetHeaders().end())
-            {
-              const std::string& content_range = content_range_iterator->second;
-              auto slash_pos = content_range.find("/");
-              result.FileSize = std::stoll(content_range.substr(slash_pos + 1));
-            }
-            else
-            {
-              result.FileSize = std::stoll(response.GetHeaders().at(Details::HeaderContentLength));
             }
             if (response.GetHeaders().find("content-type") != response.GetHeaders().end())
             {